telegram_monitor_bot/requirements.txt
python-telegram-bot>=20.0
loguru>=0.7.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
aiofiles==23.2.1
pydantic==2.5.3 